requirements.txt
streamlit>=1.24.0
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.21.0
python-dotenv>=0.19.0
pathlib>=1.0.1
tiktoken>=0.5.0
pandas>=1.3.0